Query the schema as you need to, rather than all at the start.
Previously, the entire schema was populated, recursing even
through options the user hadn't activated. This meant that
as the schema grew (and grew), the load time for diamond
crept up (and probably at a nonlinear rate). 

Now, we stop recursing at inactive options. This means that
only the elements of the tree that are actually currently
needed are queried when you open diamond: either the default
active options underneath the root (if you don't open a file),
or the options in a file (if you do).

This makes diamond much faster to start. Previously,
it used to take 30-40 seconds on my machine just to load
a file. Now it opens more or less instantaneously, provided
you have python + pygtk in your disk cache. (I note that this
strategy wasn't a problem when diamond was written: the schema
was so much smaller that it didn't matter. But now it does.)

The cost of this is that find no longer works as you
expect, as the tree elements you want to search through don't actually
exist yet. I've disabled find for now until James and I come up
with a better way to do this.

Also: fix a very silly bug that was about half of the opening
time. The logic for opening a file used to look something like:

  s = open_schema(schemafile)
  tree = s.populate_tree()
  # ... wait for 20 seconds while we query every element
  if filename is not None:
    tree = s.read_file(filename)
    # ... wait another 20 seconds as we recurse through every element, but
    # taking data from the XML file, clobbering the previous tree in the process

which was just stupid. (To be fair to us, the fact that it was doing
this was not obvious. The logic is a lot more complicated than the
pseudocode above.) Now we only populate the tree from the schema
if we don't have a filename to open.

This is quite a big change to Diamond's behaviour. I believe I've
changed the code at every place which now requires population
of the tree (e.g., when you activate a previously inactive
optional instance). But I might have missed some. I'd like
this to get some testing before the next release.