Diamond crashed ungracefully when it came across folders in the schema search paths.  Even if there   were valid schemas present, Diamond would crash.  In particular this occured when ~/.diamond/         schemata/ is in a subversion repository and there are pesky .svn folders around.
Problematic items in the search paths are now skipped and Diamond is allowed to continue to load      uninterrupted.

I've also added a note to the usage information regarding the local homespace search path.  This is a useful feature to be reminded about - particularly when you're on a train with no internet connection to automagically download a schema from the AMCG servers.

Can this change be propagated to the package repositories please?