Some improvements to diamond:
a) Trap when saving a file fails. A user in the training course
tried to write to a file she didn't have permission for. Diamond
should handle this gracefully, rather than hanging.

b) Turn off the tooltips. I've observed they just tend to get in
people's way. If you want to see the documentation, just click on it.

c) Turned on matching brackets and highlighting the current line in
the python code editor.

d) Added another datatype that the schema knows about; Simon M. came
across a schema that wanted a boolean datatype.