Modifying diamond and the spud_base schema to be able to handle syntax highlighting of different languages.
The type for python_code is now "code" with another attribute language listing it as "python".

The check for is_python_code() in tree and mixedtree is now is_code().  In both of these functions I couldn't see that:
    try:
       lang = self.get_attr("language")
       if lang == "python":
         return True
    except:
      pass
was doing anything (there was no language attribute in spud_base - until now) so I removed it.  I'm probably missing 
something so this will probably have to be incorporated back in - with some resolution between the new language attribute and
whatever this one is.

Adding a get_code_language() function to mixedtree, which returns the new attribute language.  It does a lot of defaulting to python
to try to minimize disruption to the current behaviour.

Datawidget add_text_view now retrieves the code language from the mixedtree and if it is an acceptable language specification sets
it up as the syntax highlighting.  If it's not suitable it defaults to python again.