Found a major bug in Diamond's interface, pointed out by Dan Barker.
In the MixedTree class (used to hide real_vector and stuff like that
from the user), the attributes came solely from the parent. But the
child (real_vector) element also could have attributes the user must
set -- but these were never displayed, and could never be set.

To test this, take a look at the first.rng included with this commit,
with Diamond before and after this change.