Fix exception handling for set_option through the python interface.
Previously it would raise an exception but not return the right code for it to be caught.  Now SpudNewKeyWarning can be caught in
python.

Additionally it was not possible to set a new option with type double through python.  This appears to be because the PyArg_Parse
function is broken (and its use discouraged in the documentation).  Switching this to PyFloat_AS_DOUBLE, which works.

Modifying the python tests to reflect these changes.

Also tidying up messages associated with python exceptions.