A fix for the location of the logofile.
If you've only installed from source and have multiple available schema files then you'll get an error when trying to open the
schema decision radio dialog.  This is because the logofile is None but referenced as if it were a list (logofile[0]).  This adds
the path of the source-installed logofile to the list of possible search directories.