An attempt at upgrading diamond to gtk3 and python3.
This works under both python 2 and 3 in tests so far but it does require that pygtk (and associated libraries) be replaced by pygobject with python 2 support.

Untested (and therefore definitely broken) so far are plugins and schematron and xml validation.  I don't have any examples of
plugins and the validation stuff appears broken in an unchanged build as well.