Make sure the commentwidget has an interacted attribute and that it is set to false after an update.
This hopefully fixes the bug where comments are silently dropped when the comment widget loses focus by clicking on the tree.  This
resulted in store being called twice (once through update due to the tree being clicked and once due to the comment losing focus).
On the second call the comment would be overwritten by an empty string, that was set on the first call through update.  Guaranteeing
that interacted=False after update avoids this and store does nothing on the second call.