Switch to python3 everywhere
Previously diamond and spud had been made python3 compatible. But in a
systems that keep python -> python2 (which will be all Debian/Ubuntu systems
for the foreseeable future), this doesn't actually switch over to
python3 (unless you corrupt your system).

Changes:
- change all #!/usr/bin/env to python3 instead of python
- remove unnecessary #!/usr/bin/env/python in python files that function
as a module only
- change dependencies in debian package to remove any depency on python
(which pulls in python2)
- also fix the gtk dependencies (we switched over to using python-gi
instead of python-gtk2 but this had never been updated in the control
files and the documentation)
- rename python-dxdiff and python-spud packages to python3-dxdiff and
python3-spud
- remove config.guess and config.sub
- remove hideous search and replace of #!s (to be discussed)