Remove the PangoBuffer class that was derived from a gtk TextBuffer.
This was previously used to parse markup into a TextBuffer but casting the pango Attributes is no longer possible and parsing markup
is now possible directly in a TextBuffer rendering the class unnecessary.

Also updating copyright notices and fixing documentation tags for the example ballistics schema that I noticed while debugging these
changes.