Pass schemafile (name) directly to lxml etree parser
instead of opening the file ourselves and passing it to the parser. The
latter seems to sometimes (inside fluidity jenkins container for bionic
but not on a bionic desktop) give trouble opening op a schema .rng
 where it tries to first interpret the given file object as the filename
and then objects to there being utf-8 (despite the encoding being
correctly specified in the xml header). In the documentation it actually
says it's better (faster) to pass the filename rather than the fileobject
(presumably because of the above silly buggers).

Before the fix there was a special case of the form

    if 'http' in schemafile:

which was

a) a rather unsafe way to check whether schemafile is a url
b) it turns out the etree parser is perfectly capable of opening that
url, so we don't need to.

Let's hope this doesn't break elsewhere!