Fix debian build with python 3.10 and deprecate distutils
Summary is that distutils has been deprecated for a while and will be
removed in 3.12. setuptools is its replacement. However its use as an
installation mechanism, e.g. *calling* setup.py install with CLI
arguments is also being deprecated - with no good alternative (yet)
being available. Since the whole world still depends on it I don't
expect the latter to go away any time soon though, but it does seem to
no longer care about fixing things for that frontend
which is what now cause prefix arguments to not work as expected
(installing things in /usr/local/ whilst specifying --prefix=/usr),
which breaks the building of Debian packages. What Debian/Ubuntu seem
to have done is patch setuptools to make it keep on behaving as they want,
but only if you provide --install-layout=deb. So now providing that
argument for all calls of "python setup.py install" in the Debian build.

This would have also worked in combination with distutils, but taking
this opportunity to upgrade to setuptools (since we'll have to soon
anyway). These are changes by @Patol75, taken from
https://github.com/FluidityProject/fluidity/pull/351.