Rewrite Debian/Ubuntu packaging
To avoid having to deal with the setup.py install vagueries when installing python for Debian packaging, let pybuild take care of the building and installing of python. Rewrite debian/rules to make use of debhelper for as much as possible and minimize hackery. Also remove Debian-packaging-specific hacks in Makefile.in

The recipe we are now following is to install everything in debian/tmp (instead of straight into the debian/<package>/ directories), and divide the resulting files up over the packages using debian/<package.install> files. Try to clean up some Lintin errors and warnings in the process.

Note that we are still using distutils in the setup.py files and calling
"python setup.py install" which are both deprecated. Only this
combination appears to not be hit by https://github.com/pypa/pip/issues/10978
so just leaving as is.

Also introduces:
* testing on both Focal and Jammy
* fix intermittent test failure in build process